💄 style: Fix typo
